backend/requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.12.0
nltk==3.8.1
gensim==4.3.2
sentence-transformers==3.3.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
flask==3.0.2
flask-cors==4.0.0
pyswarms==1.3.0
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.66.1
joblib==1.3.2
huggingface-hub==0.26.0
transformers==4.46.0
accelerate==0.29.3
tokenizers==0.19.1
torch==2.2.1
torchvision==0.17.1
pyarrow==23.0.1
s3fs==2026.3.0
boto3==1.42.84